requirements.txt
groq
pydantic
streamlit
python-dotenv
fastapi 
uvicorn 
httpx
python-multipart
streamlit-audiorecorder
streamlit
